Delete harvard-university-of-worcester.csl
I work for the University of Worcester, and we are no longer using this referencing style. We would like to delete/archive it to minimize confusion for our students, who are still finding it in Mendeley/Zotero/reference management software. Apologies for any breach of etiquette - I helped Chris Hill set it up, but didn't have the coding experience to follow through.